requirements.txt
nltk
python-dotenv
openai
bcrypt
rich